backend/requirements.txt
Flask
Flask-Cors
azure-ai-formrecognizer
azure-ai-textanalytics==5.2.0
python-dotenv
waitress